Moved the schema logic into a separate partial class; implemented procedure and view column schemas.
